requirements.txt
pysingleton==0.2.1
colorful==0.4.4
docopt==0.6.2
ipython==7.9.0
tag-expressions>=1.0.0
lxml==4.9.1
parse_type==0.6.0
coverage==5.5
PyYAML==5.3.1
humanize==2.6.0
